chore: [분리] 파일 클래스별로 분리
하나의 파일로 뭉쳐있던 main.swift를 4개로 분리했음
